requirements.txt
flask>=2.3.3
gunicorn>=20.1.0
requests
openai
waitress
python-dotenv



